+++ crates/usvg/src/parser/marker.rs
@@ -122,9 +122,9 @@ fn resolve(
 
         let mut clip_path = ClipPath::empty(cache.gen_clip_path_id());
 
-        let mut path = Path::with_path(Rc::new(tiny_skia_path::PathBuilder::from_rect(
+        let mut path = Path::new_simple(Rc::new(tiny_skia_path::PathBuilder::from_rect(
             clip_rect.to_rect(),
-        )));
+        )))?;
         path.fill = Some(Fill::default());
 
         clip_path.root.children.push(Node::Path(Box::new(path)));